Refactor: 셀렉션 스팟 추가 Zustand Store로 관리
- useState 대신 Zustand로 상태 관리
